start switching CLI
